% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMNLFA_final.R
\name{aMNLFA.final}
\alias{aMNLFA.final}
\title{aMNLFA final model fitting function}
\usage{
aMNLFA.final(input.object)
}
\arguments{
\item{input.object}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\description{
This function generates the final aMNLFA model on which scores will be based. Can only be run after a model containing all impact and DIF effects -- i.e., after the aMNLFA.simultaneous function.
}
\examples{
aMNLFA.final()
}
\keyword{MNLFA}

