% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMNLFA_scores.R
\name{aMNLFA.scores}
\alias{aMNLFA.scores}
\title{aMNLFA score generating function}
\usage{
aMNLFA.scores(input.object)
}
\arguments{
\item{input.object}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\description{
This function creates scores generated using aMNLFA. Can only be run after the final model has been fit -- i.e., after the aMNLFA.final function.
}
\examples{

 wd <- system.file("examplefiles",package="aMNLFA")
 ob <- aMNLFA::aMNLFA.object(dir          = wd,
                           mrdata        = xstudy,
                           indicators    = paste0("BIN_", 1:12), 
                           catindicators = paste0("BIN_", 1:12), 
                           meanimpact    = c("AGE", "GENDER", "STUDY"), 
                           varimpact     = c("AGE", "GENDER", "STUDY"), 
                           measinvar     = c("AGE", "GENDER", "STUDY"), 
                           factors       = c("GENDER", "STUDY"), 
                           ID            = "ID", 
                           thresholds    = FALSE)
 aMNLFA.scores(ob)

}
\keyword{MNLFA}

