% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{write.inp.file}
\alias{write.inp.file}
\title{Function for generating Mplus input files.}
\usage{
write.inp.file(df, outfile)
}
\arguments{
\item{df}{The data file to use.}

\item{outfile}{The output Mplus script.}
}
\description{
This is a helper function which the user will not interface with; it is used to create Mplus inputs.
}
\examples{
write.mplus.script()
}
\keyword{MNLFA}

