% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMNLFA_initial.R
\name{aMNLFA.initial}
\alias{aMNLFA.initial}
\title{aMNLFA initial model fitting function}
\usage{
aMNLFA.initial(input.object)
}
\arguments{
\item{input.object}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\description{
This function generates the initial itemwise aMNLFA models.
}
\examples{
aMNLFA.initial()
}
\keyword{MNLFA}

