% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMNLFA_itemplots.R
\name{aMNLFA.itemplots}
\alias{aMNLFA.itemplots}
\title{aMNLFA item plotting function}
\usage{
aMNLFA.itemplots(input.object)
}
\arguments{
\item{input.object}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\description{
This function generates plots of item endorsement by time, and by each covariate. This is necessary for determining which covariates to use in the MNLFA.
}
\examples{
\dontrun{
 wd <- "./aMNLFA/data"
 aMNLFA.itemplots(ob)
}
}
\keyword{MNLFA}

