% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMNLFA_scoreplots.R
\name{aMNLFA.scoreplots}
\alias{aMNLFA.scoreplots}
\title{aMNLFA score plotting function}
\usage{
aMNLFA.scoreplots(input.object)
}
\arguments{
\item{input.object}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\description{
This function creates plots of scores generated using aMNLFA. Can only be run after the aMNLFA.scores function.
}
\examples{
 wd <- tempdir()
 first<-paste0(system.file(package='aMNLFA'),"/examplefiles")
 the.list <- list.files(first,full.names=TRUE)
 file.copy(the.list,wd,overwrite=TRUE)
 
 ob <- aMNLFA::aMNLFA.object(dir = wd, 
 mrdata = xstudy, 
 indicators = paste0("BIN_", 1:12),
 catindicators = paste0("BIN_", 1:12), 
 meanimpact = c("AGE", "GENDER", "STUDY"), 
 varimpact = c("AGE", "GENDER", "STUDY"), 
 measinvar = c("AGE", "GENDER", "STUDY"),
 factors = c("GENDER", "STUDY"),
 ID = "ID",
 thresholds = FALSE)
 
 aMNLFA.scoreplots(ob)
}
\keyword{MNLFA}
