% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMNLFA_scoreplots.R
\name{aMNLFA.scoreplots}
\alias{aMNLFA.scoreplots}
\title{aMNLFA score plotting function}
\usage{
aMNLFA.scoreplots(input.object)
}
\arguments{
\item{input.object}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\description{
This function creates plots of scores generated using aMNLFA. Can only be run after the aMNLFA.scores function.
}
\examples{
\dontrun{
 wd <- "./aMNLFA/data"
 aMNLFA.scoreplots(ob)
}
}
\keyword{MNLFA}

