% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMNLFA_scores.R
\name{aMNLFA.scores}
\alias{aMNLFA.scores}
\title{aMNLFA score generating function}
\usage{
aMNLFA.scores(input.object)
}
\arguments{
\item{input.object}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\description{
This function creates scores generated using aMNLFA. Can only be run after the final model has been fit -- i.e., after the aMNLFA.final function.
}
\examples{
aMNLFA.scores()
}
\keyword{MNLFA}

