% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMNLFA_simultaneous.R
\name{aMNLFA.simultaneous}
\alias{aMNLFA.simultaneous}
\title{aMNLFA simultaneous model fitting function}
\usage{
aMNLFA.simultaneous(input.object)
}
\arguments{
\item{input.object}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\description{
This function generates the simultaneous aMNLFA model from all the initial inputs.
}
\examples{
aMNLFA.simultaneous()
}
\keyword{MNLFA}

