% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixPath.R
\name{fixPath}
\alias{fixPath}
\title{helper function - removes the final slash at the end of a given string}
\usage{
fixPath(somecharacter)
}
\arguments{
\item{somecharacter}{The aMNLFA object (created using the aMNLFA.object function) which provides instructions for the function.}
}
\value{
outcharacter - string with slash at the end deleted
}
\description{
This function generates the initial itemwise aMNLFA models.
}
\examples{
 wd <- tempdir()
 first<-paste0(system.file(package='aMNLFA'),"/examplefiles")
 fixPath(first)
}
\keyword{MNLFA}
