% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_inp_file.R
\name{write.inp.file}
\alias{write.inp.file}
\title{helper function for writing out Mplus inputs}
\usage{
write.inp.file(df, outfile)
}
\arguments{
\item{df}{- the Mplus code to be written out, in a data frame}

\item{outfile}{- a directory to which the Mplus code should be written}
}
\value{
outcharacter - string with slash at the end deleted
}
\description{
This function generates the initial itemwise aMNLFA models.
}
\examples{
 wd <- tempdir()
 first<-paste0(system.file(package='aMNLFA'),"/examplefiles")
 somestring <- as.data.frame("This is some text which would be written to an Mplus file.")
 write.inp.file(somestring, first)
}
\keyword{MNLFA}
